% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_data.R
\docType{data}
\name{combined_data}
\alias{combined_data}
\title{Simulated sample UK Biobank data including 30 individuals and 490 variables.}
\format{
A data frame with 30 rows and 490 variables.
}
\source{
{Machiela Lab}
}
\usage{
combined_data
}
\description{
A simulated dataset containing 30 individuals and 490 variables that can be used to test the functionality of the \code{tte} function. A detailed list of required variables are provided in the 'README_required_variables.txt' file.
}
\examples{
  head(combined_data)

}
\keyword{datasets}
