% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte.R
\name{tte}
\alias{tte}
\title{Prepare Time-to-Event data}
\usage{
tte(
  combined_data = NULL,
  cancer_of_interest_ICD10 = c(),
  prevalent_cancer_list = c(),
  prevalent_C_cancers = TRUE,
  incident_cancer_list = c(),
  remove_prevalent_cancer = FALSE,
  remove_self_reported_cancer = FALSE
)
}
\arguments{
\item{combined_data}{Optional. A data frame containing all specified variables. If \code{NULL}, \code{tte} assumes individual data sets have been correctly generated in a user-specified working directory.}

\item{cancer_of_interest_ICD10}{Character list specifying the ICD-10 code(s) of the cancer(s) or disease(s) of interest. A value must be specified.}

\item{prevalent_cancer_list}{Character list specifying the ICD-10 code(s) of prevalent cancer(s) or disease(s) to identify. Can be left empty ex. "c()".}

\item{prevalent_C_cancers}{Logical. If \code{TRUE}, will automatically include all ICD-10 codes that begin with 'C' to the \code{prevalent_cancer_list}. If \code{FALSE} (the default), will not include ICD-10 codes that begin with 'C' to the \code{prevalent_cancer_list}.}

\item{incident_cancer_list}{Character list specifying the ICD-10 code(s) of incident cancer(s) or disease(s) to identify. Can be left empty ex. "c()".}

\item{remove_prevalent_cancer}{Logical. If \code{TRUE}, will exclude cancer(s) or disease(s) specified in \code{prevalent_cancer_list} and/or \code{prevalent_C_cancers}. If \code{FALSE} (the default), will not exclude prevalent cancer(s) or disease(s).}

\item{remove_self_reported_cancer}{Logical. If \code{TRUE}, will remove individuals who self-reported as having cancer. If \code{FALSE} (the default), will not exclude individuals with self-reported cancer.}
}
\value{
An object of class "list". This is a named list with the following columns:

\describe{
\item{\code{f.eid}}{UK Biobank generated ID.}
\item{\code{enrollment_date}}{Individual level UK Biobank enrollment date extracted during data cleaning.}
\item{\code{prevalent_cancer_date}}{Individual level UK Biobank prevalent cancer date extracted during data cleaning. Prevalent cancer codes are user-specified within \code{prevalent_cancer_list} and \code{prevalent_C_cancers}.}
\item{\code{prevalent_cancer_type}}{Individual level UK Biobank prevalent cancer ICD-10 code extracted during data cleaning. Prevalent cancer codes are user-specified within \code{prevalent_cancer_list} and \code{prevalent_C_cancers}.}
\item{\code{prevalent_cancer}}{Prevalent cancer or disease status. Prevalent cancer codes are user-specified within \code{prevalent_cancer_list} and \code{prevalent_C_cancers}.}
\item{\code{self_reported_cancer}}{Individual self reported as having cancer to UK Biobank.}
\item{\code{death_date}}{Individual level UK Biobank death date extracted during data cleaning.}
\item{\code{attrition_date}}{Individual level UK Biobank attrition (lost to follow-up) date extracted during data cleaning.}
\item{\code{case_date}}{Date of first cancer or disease of interest ICD-10 code. Codes are user-specified within \code{cancer_of_interest_ICD10}.}
\item{\code{case_type}}{Individual level UK Biobank cancer or disease of interest ICD-10 code extracted during data cleaning. Codes are user-specified within \code{cancer_of_interest_ICD10}.}
\item{\code{case_control}}{Cancer or disease of interest status. Incident cancer codes are user-specified within \code{cancer_of_interest_ICD10}.}
\item{\code{incident_date}}{Individual level UK Biobank incident cancer date extracted during data cleaning. Incident cancer codes are user-specified within \code{incident_cancer_list}.}
\item{\code{incident_type}}{Individual level UK Biobank incident cancer ICD-10 code extracted during data cleaning. Incident cancer codes are user-specified within \code{incident_cancer_list}.}
\item{\code{incident_cancer}}{Incident cancer or disease status. Incident cancer codes are user-specified within \code{incident_cancer_list}.}
\item{\code{censor_type_cancer_control}}{Study endpoint utilized for final time-to-event date. This endpoint controls for user specified incident cancer(s) or disease(s). Incident cancer codes are user-specified within \code{incident_cancer_list}.}
\item{\code{censor_date_cancer_control}}{Study endpoint date corresponding to \code{censor_type_cancer_control} endpoint. This endpoint controls for user specified incident cancer(s) or disease(s). Incident cancer codes are user-specified within \code{incident_cancer_list}.}
\item{\code{censor_days_cancer_control}}{Final time-to-event time in days. Calculated as the difference between \code{censor_date_cancer_control} and \code{enrollment_date}. This endpoint controls for user-specified incident cancer(s) or disease(s). Incident cancer codes are user-specified within \code{incident_cancer_list}.}
\item{\code{case_control_cancer_control}}{Final time-to-event status variable. This endpoint controls for user-specified incident cancer(s) or disease(s). Incident cancer codes are user-specified within \code{incident_cancer_list}.}
\item{\code{censor_type_cancer_ignore}}{Study endpoint utilized for final time-to-event date. This endpoint ignores any user-specified incident cancer(s) or disease(s).}
\item{\code{censor_date_cancer_ignore}}{Study endpoint date corresponding to \code{censor_type_cancer_ignore} endpoint. This endpoint ignores any user-specified incident cancer(s) or disease(s).}
\item{\code{censor_days_cancer_ignore}}{Final time-to-event time in days. Calculated as the difference between \code{censor_date_cancer_ignore} and \code{enrollment_date}. This endpoint ignores any user-specified incident cancer(s) or disease(s).}
\item{\code{case_control_cancer_ignore}}{Final time-to-event status variable. This endpoint ignores any user-specified incident cancer(s) or disease(s).}
}
}
\description{
Prepares time-to-event data from raw UK Biobank \url{https://www.ukbiobank.ac.uk/} electronic medical record data.
}
\details{
This function will read-in raw UK Biobank data in 'RDS' format and generate time-to-event status and time based on user-specified parameters.
}
\examples{
# Load combined_data.rda from UKBBcleanR
mydata <- combined_data
# Set function parameters
cancer_outcome <- c("C911") # ICD-10 code of the cancer of interest
prevalent_cancers <- c("D37", "D38") # Prevalent ICD-10 codes to exclude
incident_cancers <- c("C900") # Incident cancer to control for in your analysis
example_output <- tte(
  combined_data = mydata,
  cancer_of_interest_ICD10 = cancer_outcome,
  prevalent_cancer_list = prevalent_cancers,
  prevalent_C_cancers = TRUE,
  incident_cancer_list = incident_cancers,
  remove_prevalent_cancer = FALSE,
  remove_self_reported_cancer = FALSE
)

}
